Remove ability to reuse builder on Rust side, but make it reusable in C API.
Since HTML rewriter can require a lot of parameters for the initialization we’ve used the builder pattern to construct it both in JS API and Rust code. Essentially builder contains the data that should flow into rewriter upon construction. This can be easily expressed in Rust by consuming the builder in the construction method. However, there is no natural way to express that in JS: after the construction of a rewriter builder object is still available and only solution here is to throw an error if one tries to construct one more rewriter from it. Such behavior looks quite unconventional in JS land.

I’ve noticed that when Harris asked me to provide an explicit way to free builder in C API as Rust’s builder consumption approach didn’t play well with C++’s RAII model. Therefore, in  https://bitbucket.cfdata.org/projects/ECO/repos/cool-thing/pull-requests/7/overview I’ve introduced the ability to reuse builder to construct multiple rewriters from it. The builder holds two types of objects during construction which later used by rewrite during rewritable units handlers and selector matching VM AST which references (via indices in handler vectors, not actual memory addresses) rewritable units handlers to call appropriate handlers during selector matching.  During the rewriter construction and compilation of selector program, the only bits of AST that need to be shared are handler indices. This is done by a simple reference counting pointer. The tricky part is handlers.

In Rust, closures are fat function pointers which contain a static function pointer and reference to all the data captured by a closure. Therefore, closures are the first-class object with all the memory ownership rules applicable to them. We use FnMut as a handler type which means that this closure can modify captured variables. This also means that in Rust memory model there can be only one lexical reference to this closure at a time. So, in order to share such closure, we need to use single-threaded runtime read-write lock wrapped into reference counter (Rc<RefCell<…>>).

This approach worked pretty well for a single-threaded environment, though it had some limitations - reentrant invocation of rewriter was forbidden: if you initiate rewriting from one of the handlers the newly created rewriter may try to acquire a lock for the handler from which the rewriting has been initiated and in this case single-threaded read-write lock will panic.

But there was more to it. Harris pointed me to the fact that in Workers rewriters produced from the same builder will be dispatched to the different threads and used there. This means that our approach won’t work anymore - we need atomic reference counters and multi-threaded read-write lock (i.e. mutex). Well, it wouldn’t be a hard change to implement, but then I realized that it won’t work for us either.
Let’s say that one of the handlers is async on JS side and it returns a promise that is never fulfilled. This means that thread in which handler is executed will be suspended forever and, thus, handler’s read-write lock will be locked forever, effectively blocking all other rewriters produced from the same builder.

So, the only correct way to implement desirable behavior is to clone handlers on each rewriter construction. It’s not trivial to implement on Rust side: we store handlers as trait objects and use dynamic dispatch for them, but cloning requires sized concrete type, so we need a wrapper trait which aware of the underlying concrete type in order to make closure clonable. Also, if you’ll think about Rust API alone, this approach looks quite unconventional and inconvenient: each construction clones provided handlers along with the captured data and you also need to take ensure that captured data is clonable which is rarely the desired behavior.

I’ve been thinking for a while of how we can come up with an API that will both usable from Rust and fit our exotic concurrency scenario (multithreading on C++ side which is synchronized via event loop). So, I’ve come to the conclusion that on Rust side we should disallow reuse of builder since handler can’t be shared in a general case. Also, construction itself is quite a cheap operation performance-wise. The only bit that requires relatively-heavy computations is parsing of CSS selectors, but this can be solved by extraction of parsing logic, so rewriter constructor accepts references to pre-parsed selectors instead of strings, so they can be re-used to construct multiple rewriters.

However, in C API we can still keep reusable rewriter approach: external handlers are just static function pointers and captured data is a user data pointer which lifetime we don’t control from the Rust side, and, thus, ownership control and synchronization can be delegated to the external code which controls the lifetime of the captured data. And the fact that selector parsing logic has been extracted from rewriter construction allows us to still have early errors for invalid selectors. And on Rust side we don’t really need builder anymore: builder approach is hardly less verbose than construction using built-in language primitives and more convenient API can be implemented using a macro (like we did for scheme definition in wire filter).