Consume selectors AST on compilation.
This removes leftovers from the time when we attempted to make rewriter builder reusable. It held selectors AST object, so we needed to pass it as a reference and share matching payload with an already compiled selector programs via Rc.

It is not a case anymore, so we can get rid of Rc for payload and related runtime checks.