PAR-48 Preserve the definition order of content handlers for their invocation
Initially I thought that we have an arbitrary order of handler execution caused by the fact that selector matching VM returns a hash set with handler locators which is unordered. I somehow forgot about the fact that matching payload only activates handlers, but doesn't trigger their imidiate execution. It is implemented so, because matching may be triggered by a tag hint when we don't have the whole start tag token yet.

So, execution happens separately and it was done sequentially. We've had a single function that was used for all possible execution scenarios (execute, execute and deactivate, execute and remove). And in order to support "execute and remove" scenario we needed to do backwards iteration. So, handlers were executed in the reverse order of their definition.

The only type of handler for which we needed "execute and remove" were end tag handlers, which are not user-defined and created dynamically by rewriter itself, so we don't care about their order of execution. Therefore, the only thing that we needed to do is just implement narrow-specialised iteration functions for each scenario.