PAR-43: refactor memory management
This change introduce a MemoryLimiter which is responsible to limit the
user memory usage.

The C-API's cool_thing_rewriter_build now accepts a size_t max_memory
parameter which defines the upper bound of memory allowed for instance.

An initial amount of memory can be specified to pre-allocate the
internal buffer.