Refactor parser (#26)
* Parser refactor: get rid of `ch` argument in actions

* Get rid of Cursor

* Use conventional test structure.

Currently Rust doesn't allow dynamic test generation, so in order to run integration tests we need to use a rip-off of Rust's test harness which provides this functionality. Previous version of this harness conflicted with the built-in test harness, so it was impossible to have standard `#[test]` functions and auto-generated tests in the same project. Therefore, we needed custom macro-based "framework" for basic unit tests and it was impossible to have unit test modules directly in `/src` files.

The new version of external harness uses `rustc_test` export name, so it doesn't conflict with built-in `test` anymore. This allows us separate heavy, long running integration tests from basic unit tests and standard `cargo test` command can be used now. Also, we need to export less internal stuff in test configuration.

* Simplify parser loop control

* Get rid of Chunk

* Fix rebase issues

* Get rid of ParsingLoopDirective::Continue

* Simplify parsing loop control flow. Optimise by using only errors for loop termination